full clean done, must choose what files to delete and DONE

+++ src/code/PeerInterface.java
@@ -7,6 +7,6 @@ public interface PeerInterface extends Remote {
     void backup (String file_path, Integer replicationDegree) throws RemoteException;
     void restore (String file_path) throws RemoteException;
     void delete (String file_path) throws RemoteException;
-    void reclaim (int maximum_space) throws  RemoteException;
+    void reclaim (long maximum_space) throws  RemoteException;
     void state () throws RemoteException;
 }
